@@ -3,7 +3,7 @@ import Rule from '../Rule'
 export default class NotBlankRule extends Rule {
   constructor (parameter) {
     super(parameter)
-    this.name = 'required'
+    this.name = 'not-blank'
   }
 
   check (value) {
