@@ -35,12 +35,13 @@ class RegisterFormType extends AbstractType
             ])
             ->add('password', PasswordType::class, [
                 'label' => false,
-                'attr' => ['placeholder' => $this->translator->trans('Password')],
-                'help' => $this->translator->trans('The password must contain at least %limit% chars', ['%limit%' => 6])
+                'attr'  => ['placeholder' => $this->translator->trans('Password')],
+                'help'  => $this->translator->trans('The password must contain at least %limit% chars',
+                    ['%limit%' => 6])
             ])
             ->add('password_confirmation', PasswordType::class, [
-                'label' => false,
-                'attr' => ['placeholder' => $this->translator->trans('Confirm your password')],
+                'label'    => false,
+                'attr'     => ['placeholder' => $this->translator->trans('Confirm your password')],
                 'with-key' => false,
             ]);
     }
@@ -49,7 +50,7 @@ class RegisterFormType extends AbstractType
     {
         $resolver->setDefaults([
             'data_class' => RegisterFormData::class,
-            'attr' => ['id' => 'register_form']
+            'attr'       => ['id' => 'register_form']
         ]);
     }
 }
